requirements.txt
streamlit>=1.24
pandas
fpdf
python-nmap
rapidfuzz
matplotlib
